Ignore mostly obscured windows based on the corner overlap allowance
If a window is only showing a small amount of it's window and the rest is covered by the focused window then don't add it as a candidate window. This is based on the corner allowance configuration setting. Allowing window segments that are smaller then the allowance causes problems with the logic, and having this limit allows the user to adjust how small of a visible corner will be considered a candidate.